# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: The user record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register or authenticate a user (Apple/Google/Guest/Email)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterInput"
      responses:
        "200":
          description: The user record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    User:
      type: object
      required:
        - id
        - deviceId
        - authProvider
        - createdAt
      properties:
        id:
          type: string
        deviceId:
          type: string
        authProvider:
          type: string
          enum: [apple, google, guest, email]
        email:
          type: string
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
    UserRegisterInput:
      type: object
      required:
        - authProvider
      properties:
        deviceId:
          type: string
          minLength: 1
        authProvider:
          type: string
          enum: [apple, google, guest, email]
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        displayName:
          type: string
    LoginInput:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...