# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserAuthProvider } from './userAuthProvider';

export interface User {
  id: string;
  deviceId: string;
  authProvider: UserAuthProvider;
  createdAt: Date;
  email?: string;
  displayName?: string;
}