# lib/api-zod/src/generated/types/userRegisterInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserRegisterInputAuthProvider } from './userRegisterInputAuthProvider';

export interface UserRegisterInput {
  /** @minLength 1 */
  deviceId?: string;
  authProvider: UserRegisterInputAuthProvider;
  /** @format email */
  email?: string;
  /** @minLength 6 */
  password?: string;
  displayName?: string;
}